style: sessionDetailBodyの構成を修正
